mv_ddr: extend range of time_to_nclk() function
Reviewed-on: http://vgitil04.il.marvell.com:8080/33721
Tested-by: Nadav Haklai <nadavh@marvell.com>
(cherry picked from commit f7829b32fcfcfbbf852123c5e0d2a02ff7ab6a03)
Reviewed-on: http://vgitil04.il.marvell.com:8080/33806

+++ mv_ddr_common.c
@@ -109,5 +109,5 @@ unsigned int ceil_div(unsigned int x, unsigned int y)
 unsigned int time_to_nclk(unsigned int t, unsigned int tclk)
 {
 	/* t & tclk parameters are in ps */
-	return (t * 1000 / tclk + 974) / 1000;
+	return ((unsigned long)t * 1000 / tclk + 974) / 1000;
 }
